+++ app.vue
@@ -1,5 +1,5 @@
 <template>
-  <NuxtLayout >
+  <NuxtLayout>
     <NuxtPage />
   </NuxtLayout>
 </template>
